% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixopt_list.R
\name{as.mixopt_list}
\alias{as.mixopt_list}
\title{Coerce to a mixopt_list}
\usage{
as.mixopt_list(x)
}
\arguments{
\item{x}{Object}
}
\value{
Object of class mixopt_list
}
\description{
Coerce to a mixopt_list
}
