% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixopt_list.R
\name{c_mixopt_list}
\alias{c_mixopt_list}
\title{Combines mixopt_list objects}
\usage{
c_mixopt_list(x, ...)
}
\arguments{
\item{x}{Object}

\item{...}{Additional objects}
}
\value{
A combined mixopt_list
}
\description{
Combines mixopt_list objects
}
\examples{
c_mixopt_list(NULL, as.mixopt_list(1:5), NULL, as.mixopt_list(letters[1:5]))
c_mixopt_list(as.mixopt_list(1:3), NULL)
}
