% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixopt_list.R
\name{is.mixopt_list}
\alias{is.mixopt_list}
\title{Checks if object is mixopt_list}
\usage{
is.mixopt_list(x)
}
\arguments{
\item{x}{Object}
}
\value{
TRUE if x has class "mixopt_list"
}
\description{
Checks if object is mixopt_list
}
