% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixopt.R
\name{mixopt_coorddesc}
\alias{mixopt_coorddesc}
\title{Title}
\usage{
mixopt_coorddesc(
  par,
  fn,
  gr = NULL,
  ...,
  method,
  max_iter = 100,
  verbose = 10,
  maxiter = 10,
  track_par = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Title
}
\examples{
mixopt_coorddesc(par=list(par_cts(2,8)), fn=function(x) {(4.5-x[[1]])^2})
mixopt_coorddesc(par=list(par_cts(2,8), par_unordered(letters[1:6])),
                 fn=function(x) {ifelse(x[[2]] == 'b', -1, 0) +(4.5-x[[1]])^2})
}
