% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{mopar_ordered}
\alias{mopar_ordered}
\title{Ordered variable parameter}
\usage{
mopar_ordered(values, start = NULL)
}
\arguments{
\item{values}{Values the parameter can take, in order}

\item{start}{Start parameter for optimization}
}
\value{
mixopt_par list
}
\description{
Ordered variable parameter
}
\examples{
mopar_ordered(c(1,3,5))
mopar_ordered(c('a','c'))
mopar_ordered(1:4)
mopar_ordered(4:1)
mopar_ordered(list('a', 2, 'c', sin))
}
