% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{mopar_unordered}
\alias{mopar_unordered}
\title{Unordered factor parameter}
\usage{
mopar_unordered(values, start = NULL)
}
\arguments{
\item{values}{Values the variable can take}

\item{start}{Start value. Chosen randomly if not given.}
}
\value{
mixopt_par list
}
\description{
Unordered factor parameter
}
\examples{
mopar_unordered(c(1,3,9))
mopar_unordered(letters)
}
