% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_track_par.R
\name{plot_track_par}
\alias{plot_track_par}
\title{Plot the traked parameters from an optimization}
\usage{
plot_track_par(out)
}
\arguments{
\item{out}{Output from mixopt}
}
\value{
Plot
}
\description{
Plot the traked parameters from an optimization
}
\examples{
1
}
