% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixopt_list.R
\name{[.mixopt_list}
\alias{[.mixopt_list}
\title{Index mixopt_list}
\usage{
\method{[}{mixopt_list}(x, i, value)
}
\arguments{
\item{x}{x}

\item{i}{i}

\item{value}{value}
}
\value{
value at index
}
\description{
Avoid standard list indexing which returns list for single index.
}
\examples{
a <- list(1,4,'c')
class(a) <- "mixopt_list"
a[3]
a[2:3]
}
