% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{verify_par}
\alias{verify_par}
\title{Verify parameters}
\usage{
verify_par(par)
}
\arguments{
\item{par}{List of parameters}
}
\value{
Nothing, raises error if not valid
}
\description{
Verify parameters
}
\examples{
verify_par(
  list(
    mopar_cts(2, 8, 6)
  )
)
}
