% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixopt.R
\name{mixopt}
\alias{mixopt}
\alias{mixopt_multistart}
\alias{mixopt_coorddesc}
\title{Mixed variable optimization using coordinate descent}
\usage{
mixopt(
  par,
  fn,
  gr = NULL,
  global = "multistart",
  local = "coorddesc",
  ...,
  method,
  verbose = 10,
  track
)

mixopt_multistart(
  par,
  fn,
  gr = NULL,
  ...,
  method,
  n0 = 20,
  n1 = 2,
  maxiter = 100,
  verbose = 10,
  track = FALSE
)

mixopt_coorddesc(
  par,
  fn,
  gr = NULL,
  ...,
  method,
  maxiter = 100,
  verbose = 10,
  track = FALSE
)
}
\arguments{
\item{par}{List of parameters}

\item{fn}{Function to evaluate}

\item{gr}{Gradient of fn}

\item{global}{Global optimization algorithm to use.
`FALSE` if you only want local optimization.}

\item{local}{Local optimization algorithm to use.}

\item{...}{Additional args}

\item{method}{Optimization method}

\item{verbose}{How much to print. 0 is none, 1 is standard,
2 is some, 3 is a lot, etc.}

\item{track}{Should it track the parameters evaluated and value?}

\item{n0}{For multistart, number of random initial points to evaluate.}

\item{n1}{For multistart, number of best starts to optimize with
gradient descent.}

\item{maxiter}{Maximum number of outer iterations}
}
\value{
List
}
\description{
Mixed variable optimization using coordinate descent
}
\examples{
mixopt_coorddesc(par=list(mopar_cts(2,8)), fn=function(x) {(4.5-x[[1]])^2})
mixopt_coorddesc(par=list(mopar_cts(2,8), mopar_unordered(letters[1:6])),
                 fn=function(x) {ifelse(x[[2]] == 'b', -1, 0) +(4.5-x[[1]])^2})
}
