% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixopt.R
\name{mixopt_coorddesc}
\alias{mixopt_coorddesc}
\title{Mixed variable optimization using coordinate descent}
\usage{
mixopt_coorddesc(
  par,
  fn,
  gr = NULL,
  ...,
  method,
  maxiter = 100,
  verbose = 10,
  track_par = FALSE
)
}
\arguments{
\item{par}{List of parameters}

\item{fn}{Function to evaluate}

\item{gr}{Gradient of fn}

\item{...}{Additional args}

\item{method}{Optimization method}

\item{maxiter}{Maximum number of outer iterations}

\item{verbose}{How much to print. 0 is none, 1 is standard,
2 is some, 3 is a lot, etc.}

\item{track_par}{Should it track the parameters evaluated?}
}
\value{
List
}
\description{
Mixed variable optimization using coordinate descent
}
\examples{
mixopt_coorddesc(par=list(mopar_cts(2,8)), fn=function(x) {(4.5-x[[1]])^2})
mixopt_coorddesc(par=list(mopar_cts(2,8), mopar_unordered(letters[1:6])),
                 fn=function(x) {ifelse(x[[2]] == 'b', -1, 0) +(4.5-x[[1]])^2})
}
