% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par.R
\name{par_ordered}
\alias{par_ordered}
\title{Ordered variable parameter}
\usage{
par_ordered(values, start = NULL)
}
\arguments{
\item{values}{Values the parameter can take, in order}

\item{start}{Start parameter for optimization}
}
\value{
mixopt_par list
}
\description{
Ordered variable parameter
}
\examples{
par_ordered(c(1,3,5))
#par_ordered(c(1,1))
par_ordered(c('a','c'))
par_ordered(1:4)
par_ordered(4:1)
par_ordered(list('a', 2, 'c', sin))
}
