% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixopt_list.R
\name{[.mixopt_list}
\alias{[.mixopt_list}
\title{Title}
\usage{
\method{[}{mixopt_list}(x, i, value)
}
\arguments{
\item{x}{x}

\item{i}{i}

\item{value}{value}
}
\value{
value
}
\description{
Title
}
\examples{
a <- list(1,4,'c')
class('a') <- "mixopt_list"
a[3]
}
